voice-agents/analytics/requirements.txt
smallestai>=4.2.0
python-dotenv
